+++ python/bindings/openravepy_kinbody.cpp
@@ -1776,27 +1776,6 @@ public:
         }
     };
 
-    class PyGrabbed
-    {
-public:
-        PyGrabbed(const RobotBase::Grabbed& grabbed, PyEnvironmentBasePtr pyenv) {
-            grabbedbody.reset(new PyKinBody(KinBodyPtr(grabbed._pgrabbedbody),pyenv));
-            linkrobot.reset(new PyLink(KinBody::LinkPtr(grabbed._plinkrobot),pyenv));
-
-            FOREACHC(it, grabbed._vNonCollidingLinks) {
-                nonCollidingLinks.append(PyLinkPtr(new PyLink(boost::const_pointer_cast<KinBody::Link>(*it),pyenv)));
-            }
-            troot = ReturnTransform(grabbed._troot);
-        }
-        virtual ~PyGrabbed() {
-        }
-
-        PyKinBodyPtr grabbedbody;
-        PyLinkPtr linkrobot;
-        boost::python::list nonCollidingLinks;
-        object troot;
-    };
-
     PyRobotBase(RobotBasePtr probot, PyEnvironmentBasePtr pyenv) : PyKinBody(probot,pyenv), _probot(probot) {
     }
     PyRobotBase(const PyRobotBase& r) : PyKinBody(r._probot,r._pyenv) {
@@ -2941,13 +2920,6 @@ void init_openravepy_kinbody()
         .def("__unicode__",&PyRobotBase::PyAttachedSensor::__unicode__)
         .def("__eq__",&PyRobotBase::PyAttachedSensor::__eq__)
         .def("__ne__",&PyRobotBase::PyAttachedSensor::__ne__)
-        ;
-
-        class_<PyRobotBase::PyGrabbed, boost::shared_ptr<PyRobotBase::PyGrabbed> >("Grabbed", DOXY_CLASS(RobotBase::Grabbed),no_init)
-        .def_readwrite("grabbedbody",&PyRobotBase::PyGrabbed::grabbedbody)
-        .def_readwrite("linkrobot",&PyRobotBase::PyGrabbed::linkrobot)
-        .def_readwrite("nonCollidingLinks",&PyRobotBase::PyGrabbed::nonCollidingLinks)
-        .def_readwrite("troot",&PyRobotBase::PyGrabbed::troot)
         ;
     }