+++ test/noseplugins/xunitmultiprocess.py
@@ -139,7 +139,7 @@ class Xunitmp(Plugin):
         parser.add_option(
             '--xunit-file', action='store',
             dest='xunit_file', metavar="FILE",
-            default=env.get('NOSE_XUNIT_FILE', 'nosetests.xml'),
+            default=env.get('NOSE_XUNI_FILE', 'nosetests.xml'),
             help=("Path to xml file to store the xunit report in. "
                   "Default is nosetests.xml in the working directory "
                   "[NOSE_XUNIT_FILE]"))